# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VideoGenerateInputAspectRatio = typeof VideoGenerateInputAspectRatio[keyof typeof VideoGenerateInputAspectRatio];


export const VideoGenerateInputAspectRatio = {
  '16:9': '16:9',
  '9:16': '9:16',
  '1:1': '1:1',
} as const;

export type VideoGenerateInputStyle = typeof VideoGenerateInputStyle[keyof typeof VideoGenerateInputStyle];


export const VideoGenerateInputStyle = {
  cinematic: 'cinematic',
  editorial: 'editorial',
  dreamlike: 'dreamlike',
  kinetic: 'kinetic',
} as const;

export interface VideoGenerateInput {
  /**
     * @minLength 8
     * @maxLength 500
     */
  prompt: string;
  aspectRatio?: VideoGenerateInputAspectRatio;
  style?: VideoGenerateInputStyle;
}

export interface VideoScene {
  id: string;
  start: number;
  end: number;
  title: string;
  caption: string;
  visual: string;
  accent: string;
}

export type VideoProjectStatus = typeof VideoProjectStatus[keyof typeof VideoProjectStatus];


export const VideoProjectStatus = {
  ready: 'ready',
} as const;

export interface VideoProject {
  id: string;
  prompt: string;
  title: string;
  aspectRatio: string;
  style: string;
  duration: number;
  createdAt: string;
  status: VideoProjectStatus;
  scenes: VideoScene[];
}

export interface ErrorResponse {
  error: string;
}