# lib/api-zod/src/generated/types/videoGenerateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoGenerateInputAspectRatio } from './videoGenerateInputAspectRatio';
import type { VideoGenerateInputStyle } from './videoGenerateInputStyle';

export interface VideoGenerateInput {
  /**
     * @minLength 8
     * @maxLength 500
     */
  prompt: string;
  aspectRatio?: VideoGenerateInputAspectRatio;
  style?: VideoGenerateInputStyle;
}