# lib/api-zod/src/generated/types/videoProject.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoProjectStatus } from './videoProjectStatus';
import type { VideoScene } from './videoScene';

export interface VideoProject {
  id: string;
  prompt: string;
  title: string;
  aspectRatio: string;
  style: string;
  duration: number;
  createdAt: string;
  status: VideoProjectStatus;
  scenes: VideoScene[];
}