# lib/api-zod/src/generated/types/videoScene.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoScene {
  id: string;
  start: number;
  end: number;
  title: string;
  caption: string;
  visual: string;
  accent: string;
}